+++ bzt/modules/reporting.py
@@ -66,7 +66,8 @@ class FinalStatus(Reporter, AggregatorListener):
             if self.parameters.get("failed-labels", False):
                 self.__report_failed_labels(self.last_sec[DataPoint.CUMULATIVE])
 
-        self.__report_duration()
+        if self.parameters.get("test-duration", True):
+                self.__report_duration()
 
     def __report_samples_count(self, summary_kpi_set):
         """